+++ packages/atomic-angular/projects/atomic-angular/src/lib/stencil-generated/components.ts
@@ -212,6 +212,28 @@ export class AtomicFieldCondition {
 }
 
 
+export declare interface AtomicFoldedResultList extends Components.AtomicFoldedResultList {}
+
+@ProxyCmp({
+  defineCustomElementFn: undefined,
+  inputs: ['childField', 'collectionField', 'density', 'fieldsToInclude', 'imageSize', 'parentField'],
+  methods: ['setRenderFunction']
+})
+@Component({
+  selector: 'atomic-folded-result-list',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  template: '<ng-content></ng-content>',
+  inputs: ['childField', 'collectionField', 'density', 'fieldsToInclude', 'imageSize', 'parentField']
+})
+export class AtomicFoldedResultList {
+  protected el: HTMLElement;
+  constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
+    c.detach();
+    this.el = r.nativeElement;
+  }
+}
+
+
 export declare interface AtomicFormatCurrency extends Components.AtomicFormatCurrency {}
 
 @ProxyCmp({
@@ -328,6 +350,27 @@ export declare interface AtomicLayoutSection extends Components.AtomicLayoutSect
   inputs: ['maxWidth', 'minWidth', 'section']
 })
 export class AtomicLayoutSection {
+  protected el: HTMLElement;
+  constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
+    c.detach();
+    this.el = r.nativeElement;
+  }
+}
+
+
+export declare interface AtomicLoadMoreChildrenResults extends Components.AtomicLoadMoreChildrenResults {}
+
+@ProxyCmp({
+  defineCustomElementFn: undefined,
+  inputs: ['label']
+})
+@Component({
+  selector: 'atomic-load-more-children-results',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  template: '<ng-content></ng-content>',
+  inputs: ['label']
+})
+export class AtomicLoadMoreChildrenResults {
   protected el: HTMLElement;
   constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
     c.detach();
@@ -631,6 +674,49 @@ export class AtomicResultBadge {
 }
 
 
+export declare interface AtomicResultChildren extends Components.AtomicResultChildren {}
+
+@ProxyCmp({
+  defineCustomElementFn: undefined,
+  inputs: ['imageSize', 'inheritTemplates', 'noResultText']
+})
+@Component({
+  selector: 'atomic-result-children',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  template: '<ng-content></ng-content>',
+  inputs: ['imageSize', 'inheritTemplates', 'noResultText']
+})
+export class AtomicResultChildren {
+  protected el: HTMLElement;
+  constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
+    c.detach();
+    this.el = r.nativeElement;
+  }
+}
+
+
+export declare interface AtomicResultChildrenTemplate extends Components.AtomicResultChildrenTemplate {}
+
+@ProxyCmp({
+  defineCustomElementFn: undefined,
+  inputs: ['conditions'],
+  methods: ['getTemplate']
+})
+@Component({
+  selector: 'atomic-result-children-template',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  template: '<ng-content></ng-content>',
+  inputs: ['conditions']
+})
+export class AtomicResultChildrenTemplate {
+  protected el: HTMLElement;
+  constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
+    c.detach();
+    this.el = r.nativeElement;
+  }
+}
+
+
 export declare interface AtomicResultDate extends Components.AtomicResultDate {}
 
 @ProxyCmp({
@@ -895,6 +981,25 @@ export class AtomicResultSectionBottomMetadata {
 }
 
 
+export declare interface AtomicResultSectionChildren extends Components.AtomicResultSectionChildren {}
+
+@ProxyCmp({
+  defineCustomElementFn: undefined
+})
+@Component({
+  selector: 'atomic-result-section-children',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  template: '<ng-content></ng-content>'
+})
+export class AtomicResultSectionChildren {
+  protected el: HTMLElement;
+  constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
+    c.detach();
+    this.el = r.nativeElement;
+  }
+}
+
+
 export declare interface AtomicResultSectionEmphasized extends Components.AtomicResultSectionEmphasized {}
 
 @ProxyCmp({
@@ -1060,13 +1165,13 @@ export declare interface AtomicSearchBox extends Components.AtomicSearchBox {}
 
 @ProxyCmp({
   defineCustomElementFn: undefined,
-  inputs: ['numberOfQueries', 'redirectionUrl']
+  inputs: ['numberOfQueries', 'redirectionUrl', 'suggestionTimeout']
 })
 @Component({
   selector: 'atomic-search-box',
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: '<ng-content></ng-content>',
-  inputs: ['numberOfQueries', 'redirectionUrl']
+  inputs: ['numberOfQueries', 'redirectionUrl', 'suggestionTimeout']
 })
 export class AtomicSearchBox {
   protected el: HTMLElement;