+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.10](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.13.9...@coveo/quantic@1.13.10) (2022-04-28)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
 ## [1.13.9](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.13.8...@coveo/quantic@1.13.9) (2022-04-26)